fix os.utimes/symlink failing to raise OSErrors fixes #1755

+++ Lib/test/test_os_jy.py
@@ -12,10 +12,15 @@ class OSTestCase(unittest.TestCase):
         open(test_support.TESTFN, 'w').close()
 
     def tearDown(self):
-        os.remove(test_support.TESTFN)
+        if os.path.exists(test_support.TESTFN):
+            os.remove(test_support.TESTFN)
 
     def test_issue1727(self):
         os.stat(*(test_support.TESTFN,))
+
+    def test_issue1755(self):
+        os.remove(test_support.TESTFN)
+        self.assertRaises(OSError, os.utime, test_support.TESTFN, None)
 
 
 def test_main():
